requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==0.27.8
python-dotenv==0.19.0
Werkzeug==2.3.7
